Add two new datetime-from-string functions
Using strptime is slow. datetime.fromisoformat() is fast, but limited in
what it can parse. This commit adds a new function that tries to use
isoformat, and falls back on strptime if needed, and another new
function that uses more custom code to parse timestamps.

All 3 functions pass the tests, and here are the timeit values:

2021-03-23T06:20:30.522328-04:00
  utc_datetime_from_string: 0.16827842600469012
  utc_datetime_from_string_isoformat: 0.003353077991050668
  utc_datetime_from_string_custom: 0.06806577798852231
2021-03-23T06:20:30.52232-04:00
  utc_datetime_from_string: 0.15312162099871784
  utc_datetime_from_string_isoformat: 0.1722432920068968
  utc_datetime_from_string_custom: 0.06483625598775689
2020-11-25T20:30:43.73776Z
  utc_datetime_from_string: 0.10550696100108325
  utc_datetime_from_string_isoformat: 0.1303601539984811
  utc_datetime_from_string_custom: 0.03477333100454416

Notes:
  - strptime is consistently slow
  - isoformat is very fast, but has to fall back to a slow strptime
  - the custom function is consistently much faster than strptime